Move Frame label into FrameSlider
