+++ visualizer/src/Controls/TrackingControls.js
@@ -1,0 +1,17 @@
+import { useLabeled } from '../ProjectContext';
+import UndoRedo from './Segment/UndoRedo';
+import DivisionAlerts from './Tracking/Alerts/DivisionAlerts';
+import Timeline from './Tracking/Timeline';
+
+function TrackingControls() {
+  const labeled = useLabeled();
+  return (
+    <>
+      <UndoRedo />
+      <DivisionAlerts />
+      {labeled && <Timeline />}
+    </>
+  );
+}
+
+export default TrackingControls;
